requirements.txt
streamlit>=1.31.0
google-generativeai>=0.4.0
requests>=2.31.0
Pillow>=10.0.0
streamlit-image-coordinates>=0.1.6
streamlit-drawable-canvas>=0.9.3
streamlit<1.36.0
